requirements.txt
Django>=5.2,<6.0
djangorestframework>=3.16,<4.0
django-cors-headers>=4.7,<5.0
python-dotenv>=1.1,<2.0
openai>=1.0,<2.0
requests>=2.31,<3.0
pandas>=2.0,<3.0
PyPDF2>=3.0,<4.0
openpyxl>=3.1,<4.0
xlrd>=2.0,<3.0
